requirements.txt
Pillow
BeautifulSoup4
requests
pyautogui
jupyter
ffmpeg
pywin32
pyaudio
pydub
pytesseract
pygetwindow
winshell
librosa
numpy